+++ src/frontend/apps/impress/src/features/docs/doc-management/types.tsx
@@ -59,3 +59,9 @@ export interface Doc {
     versions_retrieve: boolean;
   };
 }
+
+export enum DocDefaultFilter {
+  ALL_DOCS = 'all_docs',
+  MY_DOCS = 'my_docs',
+  SHARED_WITH_ME = 'shared_with_me',
+}